refactor(form): improve folder structure
