+++ src/components/libs/Form/FormDialog.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import user from '@testing-library/user-event';
-import { FormDialog } from '.';
+import { FormDialog } from '../index';
 
 const handleClose = jest.fn();
