+++ src/components/libs/Form/index.js
@@ -1,0 +1,7 @@
+import FormDialog from './FormDialog';
+import FormFooter from './FormFooter';
+
+export {
+  FormDialog,
+  FormFooter,
+};
